rc_max_rate does not match CodecContext

+++ src/codec/encoder/encoder.rs
@@ -76,7 +76,7 @@ impl Encoder {
 
 	pub fn set_max_bit_rate(&mut self, value: usize) {
 		unsafe {
-			(*self.as_mut_ptr()).rc_max_rate = value as i64;
+			(*self.as_mut_ptr()).rc_max_rate = value as c_int;
 		}
 	}
